ghost catch pref & ui cleanup
