+++ code/__defines/nif_ch.dm
@@ -6,5 +6,6 @@
 #define NIF_SC_ALLOW_EYES			0x8
 #define NIF_SC_BACKUPS				0x10
 #define NIF_SC_PROJECTING			0x20
+#define SOULGEM_CATCHING_GHOSTS		0x200
 #define SOULGEM_ACTIVE				0x400
 #define SOULGEM_SHOW_VORE_SFX		0x800
