+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/VoreSoulcatcher.tsx
@@ -1,25 +1,27 @@
-import { toFixed } from 'common/math';
-
 import { useBackend } from '../../../backend';
 import {
   Box,
   Button,
   Dropdown,
-  Flex,
   LabeledList,
   Section,
-  Slider,
   Stack,
 } from '../../../components';
-import { bellyData, Data, DropdownEntry, soulcatcherData } from './types';
+import { abilities, bellyData, DropdownEntry, soulcatcherData } from './types';
+import { VoreAbilities } from './VoreAbilities';
+import { CatchSettings } from './VoreSoulcatcherSettings/CatchSettings';
+import { GlobalOptions } from './VoreSoulcatcherSettings/GlobalOptions';
+import { GlobalSettings } from './VoreSoulcatcherSettings/GlobalSettings';
+import { SoulOptions } from './VoreSoulcatcherSettings/SoulOptions';
 
 export const VoreSoulcatcher = (props: {
   soulcatcher: soulcatcherData | null;
   our_bellies: Required<bellyData[]> & {
     map(arg0: (belly: bellyData) => DropdownEntry): DropdownEntry[];
   };
+  abilities: abilities;
 }) => {
-  const { soulcatcher, our_bellies } = props;
+  const { soulcatcher, our_bellies, abilities } = props;
 
   const getBellies = our_bellies.map((belly) => {
     return { ...belly, displayText: belly.name, value: belly.ref };
@@ -33,7 +35,7 @@ export const VoreSoulcatcher = (props: {
           overlayBellies={getBellies}
         />
       )}
-      <VoreAbilities />
+      <VoreAbilities abilities={abilities} />
     </>
   );
 };
@@ -42,7 +44,7 @@ const VoreSoulcatcherSection = (props: {
   soulcatcher: soulcatcherData;
   overlayBellies: DropdownEntry[];
 }) => {
-  const { act } = useBackend<Data>();
+  const { act } = useBackend();
 
   const { soulcatcher, overlayBellies } = props;
 
@@ -54,6 +56,7 @@ const VoreSoulcatcherSection = (props: {
     interior_design,
     catch_self,
     catch_prey,
+    catch_ghost,
     ext_hearing,
     ext_vision,
     mind_backups,
@@ -109,186 +112,20 @@ const VoreSoulcatcherSection = (props: {
               </Stack.Item>
             </Stack>
           </LabeledList.Item>
-          {selected_soul ? (
-            <LabeledList.Item label="Soul Options">
-              <Box>
-                {!taken_over ? (
-                  <>
-                    <Button
-                      icon="key"
-                      tooltip="Release the currently selected soul as ghosts."
-                      tooltipPosition="bottom"
-                      onClick={() => act('soulcatcher_release')}
-                      selected
-                    >
-                      Release
-                    </Button>
-                    <Button
-                      icon="right-left"
-                      tooltip="Tansfer the currently selected soul into a held or nearby Sleevemate or MMI."
-                      tooltipPosition="bottom"
-                      onClick={() => act('soulcatcher_transfer')}
-                    >
-                      Transfer
-                    </Button>
-                    <Button.Confirm
-                      icon="skull"
-                      tooltip="Delete the currently selected soul if preferences align or release it."
-                      tooltipPosition="bottom"
-                      color="red"
-                      confirmIcon="triangle-exclamation"
-                      onClick={() => act('soulcatcher_delete')}
-                    >
-                      Delete
-                    </Button.Confirm>
-                  </>
-                ) : (
-                  ''
-                )}
-                <Button
-                  icon="arrows-left-right"
-                  color="yellow"
-                  tooltip="Transfer control to the selected soul."
-                  tooltipPosition="bottom"
-                  onClick={() => act('soulcatcher_transfer_control')}
-                >
-                  Transfer Control
-                </Button>
-              </Box>
-            </LabeledList.Item>
-          ) : (
-            ''
-          )}
-          {caught_souls.length ? (
-            <LabeledList.Item label="Global Options">
-              <Box>
-                {!taken_over ? (
-                  <>
-                    <Button.Confirm
-                      icon="tornado"
-                      tooltip="Release all captured souls as ghosts."
-                      tooltipPosition="bottom"
-                      confirmColor="green"
-                      confirmIcon="triangle-exclamation"
-                      onClick={() => act('soulcatcher_release_all')}
-                    >
-                      Release Souls
-                    </Button.Confirm>
-                    <Button.Confirm
-                      icon="eraser"
-                      tooltip="Delete all captured souls if preferences align or release them."
-                      tooltipPosition="bottom"
-                      color="red"
-                      confirmIcon="triangle-exclamation"
-                      onClick={() => act('soulcatcher_erase_all')}
-                    >
-                      Erase Souls
-                    </Button.Confirm>
-                  </>
-                ) : (
-                  <Button
-                    icon="arrow-up-from-bracket"
-                    color="green"
-                    tooltip="Release the body back to the owner."
-                    tooltipPosition="bottom"
-                    onClick={() => act('soulcatcher_release_control')}
-                  >
-                    Release Control
-                  </Button>
-                )}
-              </Box>
-            </LabeledList.Item>
-          ) : (
-            ''
-          )}
-          <LabeledList.Item label="Global Settings">
-            <Box>
-              <Button
-                icon="circle-user"
-                tooltip={
-                  (catch_self ? 'Allow' : 'Disallow') +
-                  ' your own soulcatcher to capture your own soul upon vore death.'
-                }
-                tooltipPosition="bottom"
-                color={catch_self ? 'green' : 'red'}
-                onClick={() => act('toggle_self_catching')}
-              >
-                Catch User
-              </Button>
-              <Button
-                icon="person"
-                tooltip={
-                  (catch_prey ? 'Allow' : 'Disallow') +
-                  ' your own soulcatcher to capture prey souls upon vore death.'
-                }
-                tooltipPosition="bottom"
-                color={catch_prey ? 'green' : 'red'}
-                onClick={() => act('toggle_prey_catching')}
-              >
-                Catch Prey
-              </Button>
-              <Button
-                icon={ext_hearing ? 'ear-listen' : 'ear-deaf'}
-                tooltip={
-                  (ext_hearing ? 'Allow' : 'Disallow') +
-                  ' your captured souls to hear.'
-                }
-                tooltipPosition="bottom"
-                color={ext_hearing ? 'green' : 'red'}
-                onClick={() => act('toggle_ext_hearing')}
-              >
-                Ext. Hearing
-              </Button>
-              <Button
-                icon={ext_vision ? 'eye' : 'eye-slash'}
-                tooltip={
-                  (ext_vision ? 'Allow' : 'Disallow') +
-                  ' your captured souls to see.'
-                }
-                tooltipPosition="bottom"
-                color={ext_vision ? 'green' : 'red'}
-                onClick={() => act('toggle_ext_vision')}
-              >
-                Ext. Vision
-              </Button>
-              <Button
-                icon="database"
-                tooltip={
-                  (mind_backups ? 'Allow' : 'Disallow') +
-                  ' your captured souls to have regular mind backups.'
-                }
-                tooltipPosition="bottom"
-                color={mind_backups ? 'green' : 'red'}
-                onClick={() => act('toggle_mind_backup')}
-              >
-                Mind Backups
-              </Button>
-              <Button
-                icon="street-view"
-                tooltip={
-                  (ar_projecting ? 'Allow' : 'Disallow') +
-                  ' your captured souls to AR project themselves.'
-                }
-                tooltipPosition="bottom"
-                color={ar_projecting ? 'green' : 'red'}
-                onClick={() => act('toggle_ar_projecting')}
-              >
-                AR Projecting
-              </Button>
-              <Button
-                icon={show_vore_sfx ? 'circle-play' : 'circle-pause'}
-                tooltip={
-                  (show_vore_sfx ? 'Show' : 'Hide') +
-                  ' the selected vore SFX to your captured souls.'
-                }
-                tooltipPosition="bottom"
-                color={show_vore_sfx ? 'green' : 'red'}
-                onClick={() => act('toggle_vore_sfx')}
-              >
-                Show Vore SFX
-              </Button>
-            </Box>
-          </LabeledList.Item>
+          {selected_soul ? <SoulOptions taken_over={taken_over} /> : ''}
+          {caught_souls.length ? <GlobalOptions taken_over={taken_over} /> : ''}
+          <CatchSettings
+            catch_self={catch_self}
+            catch_prey={catch_prey}
+            catch_ghost={catch_ghost}
+          />
+          <GlobalSettings
+            ext_hearing={ext_hearing}
+            ext_vision={ext_vision}
+            mind_backups={mind_backups}
+            ar_projecting={ar_projecting}
+            show_vore_sfx={show_vore_sfx}
+          />
           <LabeledList.Item
             label="Interior Design"
             buttons={
@@ -334,54 +171,6 @@ const VoreSoulcatcherSection = (props: {
       ) : (
         'Soulcatching disabled.'
       )}
-    </Section>
-  );
-};
-
-const VoreAbilities = (props) => {
-  const { act, data } = useBackend<Data>();
-  const { nutrition, current_size, minimum_size, maximum_size, resize_cost } =
-    data.abilities;
-
-  function is_enabled(nutri: number, cost: number): boolean {
-    return nutri >= cost;
-  }
-
-  return (
-    <Section title="Abilities" buttons={'Nutrition: ' + toFixed(nutrition, 1)}>
-      <LabeledList>
-        <LabeledList.Item label="Resize">
-          <Flex align="baseline">
-            <Flex.Item>
-              <Slider
-                disabled={!is_enabled(nutrition, resize_cost)}
-                width="250px"
-                ranges={
-                  (is_enabled(nutrition, resize_cost) && {
-                    bad: [1, 25],
-                    average: [25, 50],
-                    green: [50, 150],
-                    yellow: [150, 200],
-                    red: [200, 600],
-                  }) ||
-                  (!is_enabled(nutrition, resize_cost) && { black: [0, 600] })
-                }
-                format={(value: number) => toFixed(value, 2) + '%'}
-                value={current_size * 100}
-                minValue={minimum_size * 100}
-                maxValue={maximum_size * 100}
-                onChange={(e, value: number) =>
-                  act('adjust_own_size', {
-                    new_mob_size: value / 100,
-                  })
-                }
-              />
-            </Flex.Item>
-            <Flex.Item color="label">&nbsp;&nbsp;Cost:&nbsp;</Flex.Item>
-            <Flex.Item>{resize_cost}</Flex.Item>
-          </Flex>
-        </LabeledList.Item>
-      </LabeledList>
     </Section>
   );
 };
